update playlist data
