+++ src/components/app/component.tsx
@@ -6,36 +6,9 @@ import { SocialMedia, GithubLink } from '@/components/social-media'
 import { LastUpdate } from '@/components/recent-tracks/last-update'
 import { Icons } from '@/components/icons'
 import { Mix } from '@/components/mix'
-import { m1Tracklist, Mix as MixType } from '@/data/mixes'
+import { mixes } from '@/data/mixes'
+import { playlists } from '@/data/playlists'
 import './style.scss'
-
-export const playlistIDs: string[] = [
-  'spotify:playlist:1EOj4DNc4LJD6ujTkFu6N7',
-  'spotify:playlist:0Gm4bXeS0P3YJnDyEoH3KR',
-  'spotify:playlist:1IoSlQbiD7OZ7APX3tviuX',
-  'spotify:playlist:0Qf5El4jvjpUnsJ4YZdTKo',
-
-  'https://open.spotify.com/playlist/5ywVhKummH20oqmvVDaSpm?si=d995ff38c4294b8e',
-]
-
-export const mixes: MixType[] = [{
-  url: 'https://soundcloud.com/l1fescape/moongate',
-  title: 'moongate',
-  soundcloud: '1204353763',
-  spotify: '5ywVhKummH20oqmvVDaSpm',
-}, {
-  url: 'https://soundcloud.com/l1fescape/mix1',
-  title: 'm1',
-  tracklist: m1Tracklist,
-  soundcloud: '1053816040',
-  spotify: '544H4ZhAWa94tbegE93ap',
-  description: 'i recorded a mix of a few tunes i listened to a lot in the month of may 2021.',
-}, {
-  url: 'https://soundcloud.com/l1fescape/mix2',
-  title: 'm2',
-  soundcloud: '1155198274',
-  spotify: '2VxPeV9cgCmiIAQfDouQ7T',
-}]
 
 export const App = () => (
   <div className="home">
@@ -60,8 +33,8 @@ export const App = () => (
           <h3>playlists</h3>
           <p>these are a few spotify playlists i've made with tunes i really like. check 'em out if you're looking for some music to groove to.</p>
           <ul>
-            {playlistIDs.map(id => (
-              <li key={id}><Playlist id={id} /></li>
+            {playlists.map(playlist => (
+              <li key={playlist.url}><Playlist playlist={playlist} /></li>
             ))}
           </ul>
         </div>
@@ -98,7 +71,7 @@ export const App = () => (
     <section className="source">
       <div className="title"><h2>source</h2></div>
       <div className="content">
-        <p>this site was built using</p>
+        <p>this site is built using</p>
         <ul>
           <li>frontend: typescript, react, sass, webpack</li>
           <li>ci/cd: github actions, netlify</li>
