+++ src/components/playlist/component.tsx
@@ -1,12 +1,27 @@
 import * as React from 'react'
 
-import { Spotify } from './spotify'
+import { Playlist as PlaylistType } from '@/data/playlists'
 import './style.scss'
 
 export interface PlaylistProps {
-  id: string
+  playlist: PlaylistType
 }
 
-export const Playlist: React.FC<PlaylistProps> = ({ id }) => (
-  <Spotify id={id} />
-)
+export const Playlist: React.FC<PlaylistProps> = ({ playlist }) => {
+  const { title, description, url, imageUrl } = playlist
+  return (
+    <div className="playlist">
+      <div className="art">
+        <a href={url}>
+          <img src={imageUrl} />
+        </a>
+      </div>
+      <div className="info">
+        <a href={url}>
+          <span>{title}</span>
+        </a>
+        <p>{description}</p>
+      </div>
+    </div>
+  )
+}
