--- src/components/playlist/spotify/index.ts
@@ -1,2 +1,0 @@
-export * from './component'
-export * from '../../../api/spotify/playlist'
