+++ src/data/playlist-data.json
@@ -1,1 +1,1 @@
-{"playlists":{"spotify:playlist:1EOj4DNc4LJD6ujTkFu6N7":{"title":"✨ 4am","description":"late night tunes that are great to work to","url":"https://open.spotify.com/playlist/1EOj4DNc4LJD6ujTkFu6N7","imageUrl":"https://mosaic.scdn.co/640/ab67616d0000b273235e762de339c6076aa824ddab67616d0000b273665c8dc1a815f56e6898db78ab67616d0000b2736b3a70af8e425b93a457df0cab67616d0000b2739618d5229371308305efecf1"},"https://open.spotify.com/playlist/19x8zn8rBoAzl0EEciH14w?si=1uNY_xZFQxK8iOXeIYlLfg":{"title":"🥵 it's hot","description":"sweaty beats for when it's hot on the dancefloor","url":"https://open.spotify.com/playlist/19x8zn8rBoAzl0EEciH14w","imageUrl":"https://mosaic.scdn.co/640/ab67616d0000b27329222caea52a396d250c93adab67616d0000b27332344aa2f209857499528b38ab67616d0000b27339f2257ad828ee8432641178ab67616d0000b27357a911f38f087d05f6a5a681"},"spotify:playlist:1IoSlQbiD7OZ7APX3tviuX":{"title":"🌊 floating","description":"ethereal, dreamy beats","url":"https://open.spotify.com/playlist/1IoSlQbiD7OZ7APX3tviuX","imageUrl":"https://mosaic.scdn.co/640/ab67616d0000b27304da509a0f3b1026c7a808a4ab67616d0000b2730bc26f501f81557716a4a098ab67616d0000b2739a2d7013598e55bcf699ad94ab67616d0000b273f0437a4189e1abee866b6911"},"spotify:playlist:0Qf5El4jvjpUnsJ4YZdTKo":{"title":"🎴 ether","description":"ashes and reverb","url":"https://open.spotify.com/playlist/0Qf5El4jvjpUnsJ4YZdTKo","imageUrl":"https://mosaic.scdn.co/640/ab67616d0000b2734f4e2053ad80727f313b15eeab67616d0000b27358c5865816bdc1426beeb556ab67616d0000b273613a5f8a790de7912a1aaaaeab67616d0000b273f0c38230724b54ba6eb3f088"},"spotify:playlist:0Gm4bXeS0P3YJnDyEoH3KR":{"title":"mc","description":"beats to relax/play minecraft to","url":"https://open.spotify.com/playlist/0Gm4bXeS0P3YJnDyEoH3KR","imageUrl":"https://mosaic.scdn.co/640/ab67616d0000b2734cf0b29eb06a92aa96acae64ab67616d0000b273a8534dab41e9e3b124a249b6ab67616d0000b273cbd2b9ae1cfa0b13b2fa0d1cab67616d0000b273dd0ef93e4a39e112fae7a2b1"},"spotify:playlist:5443H4ZhAWa94tbegE93ap":{"title":"m1","description":"https://soundcloud.com/l1fescape/mix1","url":"https://open.spotify.com/playlist/5443H4ZhAWa94tbegE93ap","imageUrl":"https://i.scdn.co/image/ab67706c0000bebb4461e4d020863eb9b6afee5e"}},"updated":1623775370712}
+{"playlists":{"spotify:playlist:1EOj4DNc4LJD6ujTkFu6N7":{"title":"✨ 4am","description":"late night tunes that are great to work to","url":"https://open.spotify.com/playlist/1EOj4DNc4LJD6ujTkFu6N7","imageUrl":"https://mosaic.scdn.co/640/ab67616d0000b273235e762de339c6076aa824ddab67616d0000b273665c8dc1a815f56e6898db78ab67616d0000b2736b3a70af8e425b93a457df0cab67616d0000b2739618d5229371308305efecf1"},"spotify:playlist:1IoSlQbiD7OZ7APX3tviuX":{"title":"🌊 floating","description":"ethereal, dreamy tunes","url":"https://open.spotify.com/playlist/1IoSlQbiD7OZ7APX3tviuX","imageUrl":"https://mosaic.scdn.co/640/ab67616d0000b27304da509a0f3b1026c7a808a4ab67616d0000b2739a2d7013598e55bcf699ad94ab67616d0000b273f0437a4189e1abee866b6911ab67616d0000b273f4d933675ac8f809344efad1"},"spotify:playlist:0Qf5El4jvjpUnsJ4YZdTKo":{"title":"🎴 ether","description":"ashes and reverb","url":"https://open.spotify.com/playlist/0Qf5El4jvjpUnsJ4YZdTKo","imageUrl":"https://mosaic.scdn.co/640/ab67616d0000b2734f4e2053ad80727f313b15eeab67616d0000b273508540666a650e316ffc91ceab67616d0000b27358c5865816bdc1426beeb556ab67616d0000b273613a5f8a790de7912a1aaaae"},"https://open.spotify.com/playlist/52MnZX7QxVTmL76e9eXrmK?si=4bb07212714549ea":{"title":"moongate overflow","description":"tracks that didnt make it in my moongate set","url":"https://open.spotify.com/playlist/52MnZX7QxVTmL76e9eXrmK","imageUrl":"https://mosaic.scdn.co/640/ab67616d0000b2734f95e183a8f675eed6a9a369ab67616d0000b2735b5fd439a9153812a169877eab67616d0000b2737f8d64a63ff9136a6f135099ab67616d0000b2738e2e9cd43cd0f99a2604981a"},"https://open.spotify.com/playlist/1MMqJT3c36EPJ0IYFo8DNn?si=71307ff6d9134cf4":{"title":"344bpm","description":"\"more bpm! 180bpm!!\" dj mell g/dj fuckoff inspired","url":"https://open.spotify.com/playlist/1MMqJT3c36EPJ0IYFo8DNn","imageUrl":"https://i.scdn.co/image/ab67706c0000bebbee42d9a16a64f79cc9634e9e"},"https://open.spotify.com/playlist/30iI7N8nbaTjuxO0C8cJbW?si=8c28560944a349d1":{"title":"min techno","description":"minimal techno","url":"https://open.spotify.com/playlist/30iI7N8nbaTjuxO0C8cJbW","imageUrl":"https://mosaic.scdn.co/640/ab67616d0000b2738783337e78aed82de05d46b9ab67616d0000b273892e35faa762d8bb474fba86ab67616d0000b27396acf4e52d73fabd7ac84be7ab67616d0000b273bc82591ab5f14e7ac8db863f"}},"updated":1643266652334}
